Refactor the fs interface

+++ internal/service/inodes/init.go
@@ -21,7 +21,7 @@ type Service interface {
 	HardDelete(ctx context.Context, inode uuid.UUID) error
 	CreateDir(ctx context.Context, cmd *PathCmd) (*INode, error)
 	CreateFile(ctx context.Context, cmd *CreateFileCmd) (*INode, error)
-	RegisterWrite(ctx context.Context, inode *INode, sizeWrite int, h hash.Hash) error
+	RegisterWrite(ctx context.Context, inode *INode, sizeWrite int64, h hash.Hash) error
 	GetINodeRoot(ctx context.Context, inode *INode) (*INode, error)
 	MkdirAll(ctx context.Context, cmd *PathCmd) (*INode, error)
 }
